Polytropic model, attempt 1
This commit contains
1. A change to the Chandra infile to use the correct exposure time. Background exposure time still needs correction.
2. An attempt at implementing a polytropic model. Details follow.

The polytropic model skips calculation of a number of derived  parameters until the core is more tested.
Comments beginning with TODO have been left throughtout to flag areas needing futher attention.
This particular implementation is based on Appendix C of Ghirardini et. al. 2019 (A&A 627, A19 (2019))
It assumes GammaR and Gamma0 as found in their paper are correct for all clusters, rather than sampling for them.
Prior ranges are guesstimates.